+++ HISTORY.rst
@@ -24,6 +24,7 @@ End-User Summary
 - Adding progress bar on top of case list.
 - Improving case list and detail overview page layout and usability.
 - Upgrade of the SODAR-core library app, includes various improvements such background job pagination and improvements to membership management.
+- Included tables for converting refseq and ensembl gene ids to gene symbols.
 
 Full Change List
 ================
@@ -46,6 +47,7 @@ Full Change List
 - Adding progress bar on top of case list.
 - Improving case list and detail overview page layout and usability.
 - Upgraded to SODAR core v0.7.0.
+- Included tables ``RefseqToGeneSymbol`` and ``EnsemblToGeneSymbol`` convert gene ids to gene symbols to get a better coverage of gene symbols.
 
 -------
 v0.17.4